Move JSONParser to JSONReducer of artifacts
Move JSONParser to JSONReducer file to have one
complete Workflow in one file.
