+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/Artifacts.js
@@ -3,6 +3,7 @@ import {
   REMOVE_ITEM,
   SET
 } from '../../../../../../../src/js/constants/TransactionTypes';
+import Transaction from '../../../../../../../src/js/structs/Transaction';
 import {isEmpty} from '../../../../../../../src/js/utils/ValidatorUtil';
 
 function getJson(data) {
@@ -37,6 +38,21 @@ function processTransaction(state, {type, path, value}) {
 }
 
 module.exports = {
+  JSONParser(state) {
+    if (state.fetch == null) {
+      return [];
+    }
+
+    return state.fetch.reduce(function (memo, item, index) {
+      memo.push(new Transaction( ['fetch'], index, ADD_ITEM ));
+      Object.keys(item).forEach(function (key) {
+        memo.push(new Transaction( ['fetch', index, key], item[key], SET));
+      });
+
+      return memo;
+    }, []);
+  },
+
   JSONReducer(state, {type, path, value}) {
     if (path == null) {
       return state;
